// EOD_Project/src/app/Model/combined-model.ts
export class CombinedModel {
    Id: string | any;
    WorkId?: string | any;
    Title?: string | any;
    Date: Date | any;
    AreaPath?: string | any;
    ADOTime?: string | any;
    ProactTime?: string | any;
    WorkType?: string | any;
    LeaveType?: string | any;
    Note?: string | any;
    DayType: boolean | any;
}